Let the default value of delayed be user-configurable
Before this modification, the delayed option let the user allocate the
data on CPU and GPU after the call of field_new, but the two were still
happening at the same time.
With this modification, the allocation on CPU and the allocation on GPU
are separated. To do so, the satus system had to be slighty modified
since the whole library used to assume that the two data pointers were
allocated at the same time. Finally the usege of the status system has
been simplified with the addition of three helper functions (add, remove,
is) that let us modify or query the status without having to do bits
comparisons in the middle of the code.